+++ KOF.Core/Models/Loot.cs
@@ -11,7 +11,7 @@ namespace KOF.Core.Models
     public class Loot
     {
         public uint BundleId { get; set; }
-        public short NpcId { get; set; }
+        public int NpcId { get; set; }
         public Vector3 Position { get; set; }
         public int DropTime { get; set; }
         public byte ItemCount { get; set; }
